+++ CHANGELOG.md
@@ -6,3 +6,5 @@
 2.0.0
  - Add text cleansing functions for the collected tweet
  - Unit tests added for the text cleansing functions
+ - Add the classifier, training set and unit test for the classifier
+ - Add neutral sentiment as a answer